requirements.txt
cffi
aiohttp==1.0.5
async-timeout==1.1.0
chardet==2.3.0
discord.py[voice]==0.16.7
multidict==2.1.4
olefile==0.44
Pillow==4.0.0
protobuf==3.2.0
requests==2.13.0
six==1.10.0
websockets==3.2
colour==0.1.2
wordcloud==1.3.1
youtube-dl
redis

